+++ istr/istr.py
@@ -5,7 +5,7 @@
 #    |_||___/ \__||_|
 # strings you can count on
 
-__version__ = "1.1.12"
+__version__ = "1.1.16"
 import functools
 import itertools
 import types
@@ -397,13 +397,19 @@ class istr(str):
         return not istr.is_divisible_by(self, 2)
 
     def is_divisible_by(self, divisor):
-        return istr.interpret_as_int(self) % int(divisor) == 0
+        return divisor!=0 and istr.interpret_as_int(self) % int(divisor) == 0
 
     def is_square(self):
-        return istr.is_power_of(self, 2)
+        n = istr.interpret_as_int(self)
+        if n < 1000000:
+            return n in _squares_up_to_1_000_000()
+        return n == round(n ** (1 / 2)) ** 2
 
     def is_cube(self):
-        return istr.is_power_of(self, 3)
+        n = istr.interpret_as_int(self)
+        if n < 1000000:
+            return n in _cubes_up_to_1_000_000()
+        return n == round(n ** (1 / 3)) ** 3
 
     def is_power_of(self, exponent):
         n = istr.interpret_as_int(self)
@@ -411,7 +417,7 @@ class istr(str):
             raise ValueError(f"exponent must be >=1; not {exponent}")
         if not isinstance(exponent, int):
             raise TypeError(f"exponent must be int; not {type(exponent)}")
-        return n >= 0 and self == round(n ** (1 / exponent)) ** exponent
+        return n >= 0 and n == round(n ** (1 / exponent)) ** exponent
 
     def is_prime(self):
         n = istr.interpret_as_int(self)
@@ -421,11 +427,49 @@ class istr(str):
             return True
         if not n & 1:
             return False
-
+        if n < 1000000:
+            return n in _primes_up_to_1_000_000()
         for x in range(3, int(n**0.5) + 1, 2):
             if n % x == 0:
                 return False
         return True
+
+    @classmethod
+    def primes(cls, lb_or_ub, ub=None, cache=True):
+        """
+        returns all primes up to a given upperbound or between a given lowerbound and upperbound
+        """
+        if ("primes", lb_or_ub, ub) in _cache:
+            return _cache["primes", lb_or_ub, ub]
+        result = sorted(map(cls, _primes(lb_or_ub, ub)))
+        if cache:
+            _cache["primes", lb_or_ub, ub]=result
+        return result 
+
+    @classmethod
+    def squares(cls, lb_or_ub, ub=None, cache=True):
+        """
+        returns all squares up to a given upperbound or between a given lowerbound and upperbound
+        """
+        return istr.power_ofs(2,lb_or_ub, ub, cache=cache)
+
+    @classmethod
+    def cubes(cls, lb_or_ub, ub=None, cache=True):
+        return istr.power_ofs(3,lb_or_ub, ub, cache=cache)
+
+    @classmethod
+    def power_ofs(cls, n,lb_or_ub, ub=None, cache=True):
+        """
+        returns all power of n up to a given upperbound or between a given lowerbound and upperbound
+        """
+        if ("power_ofs", n, lb_or_ub, ub) in _cache:
+            return _cache["power_ofs", n, lb_or_ub, ub]
+        result = sorted(map(cls, _power_ofs(n,lb_or_ub, ub)))
+        if cache:
+            _cache["power_ofs", n, lb_or_ub, ub]=result
+        return result         
+
+
 
     def decompose(self, letters, namespace=None):
         """
@@ -442,7 +486,7 @@ class istr(str):
             if letter in lookup and lookup[letter] != ch:
                 raise ValueError(f"multiple values found for variable {letter}")
             if not letter.isidentifier():
-                raise ValueError(f"{letter} cannot be used as a variable")
+                raise ValueError(f"{repr(letter)} cannot be used as a variable")
             lookup[str(letter)] = ch
         if len(letters) != len(self):
             raise ValueError(f"incorrect number of variables {len(letters)}; should be {len(self)}")
@@ -457,7 +501,7 @@ class istr(str):
             namespace = inspect.currentframe().f_back.f_globals
         for letter in letters:
             if letter not in namespace:
-                raise ValueError(f"variable {letter} not defined")
+                raise ValueError(f"variable {repr(letter)} not defined")
         s = "".join(str(namespace[letter]) for letter in letters)
         return istr(s)
 
@@ -504,7 +548,7 @@ class istr(str):
 
     def is_triangular(self):
         n = istr.interpret_as_int(self)
-        if n<=0:
+        if n <= 0:
             return False
         return istr.is_square(n * 8 + 1)
 
@@ -542,6 +586,17 @@ class istr(str):
 
     def is_int(self):
         return self._as_int is not self._nan
+
+    def join(self, iterable=None):
+        if isinstance(self, istr):
+            return istr(str(self).join(iterable))
+        else:
+            if iterable is None:
+                return istr("").join(self)
+            else:
+                if not isinstance(self, str):
+                    raise TypeError(f"{self!r} should be str, not {type(self)}")
+                return istr(self).join(iterable)
 
     @classmethod
     def concat(cls, iterable):
@@ -585,7 +640,8 @@ class istr(str):
 
             cls._int_format = int_format
 
-        def __enter__(self): ...
+        def __enter__(self):
+            ...
 
         def __exit__(self, exc_type, exc_value, exc_tb):
             self.saved_cls._int_format = self.saved_int_format
@@ -606,7 +662,8 @@ class istr(str):
             self.saved_cls = cls
             cls._repr_mode = mode
 
-        def __enter__(self): ...
+        def __enter__(self):
+            ...
 
         def __exit__(self, exc_type, exc_value, exc_tb):
             self.saved_cls._repr_mode = self.saved_repr_mode
@@ -625,7 +682,8 @@ class istr(str):
             self.saved_cls = cls
             cls._base = base
 
-        def __enter__(self): ...
+        def __enter__(self):
+            ...
 
         def __exit__(self, exc_type, exc_value, exc_tb):
             self.saved_cls._base = self.saved_base
@@ -738,9 +796,51 @@ def _map(func, *iterables, strict=False):
 
         yield func(*values)
 
+_cache={}
+
+def _primes(lb_or_ub, ub=None):
+    lb, ub = (0, lb_or_ub) if ub is None else (lb_or_ub, ub)
+    sieve = bytearray(b"\x01") * (ub + 1)
+    sieve[0:2] = b"\x00\x00"
+
+    for i in range(2, int(ub**0.5)+1 ):
+        if sieve[i]:
+            sieve[i * i : ub + 1 : i] = b"\x00" * (((ub - i * i) // i) + 1)
+
+    return {i for i, is_prime in enumerate(sieve) if is_prime and lb<=i<ub}
+
+
+@functools.lru_cache(maxsize=1)
+def _primes_up_to_1_000_000():
+    return _primes(1000000)
+
+
+@functools.lru_cache(maxsize=1)
+def _squares_up_to_1_000_000():
+    return _power_ofs(2,1000000)
+
+@functools.lru_cache(maxsize=1)
+def _cubes_up_to_1_000_000():
+    return _power_ofs(3,1000000)
+
+
+def _power_ofs(n, lb_or_ub, ub=None):
+    lb, ub = (0, lb_or_ub) if ub is None else (lb_or_ub, ub)
+    result = set()
+    if n==0:
+        if lb<=1<ub:
+            result.add(1)
+    elif n==1:
+        result={*range(lb,ub)}
+    else:
+        i = int(lb ** (1 / n))
+        while (in_ := i**n) < ub:
+            if in_ >= lb:
+                result.add(in_)
+            i += 1
+    return result
 
 istr.type = type(istr(0))
-
 
 class istrModule(types.ModuleType):
     def __call__(self, *args, **kwargs):
@@ -755,3 +855,4 @@ class istrModule(types.ModuleType):
 
 if __name__ != "__main__":
     sys.modules["istr"].__class__ = istrModule
+